Added support for listing all activity for Application, Module and Home domains.
Fixed bug in ActivityTraceService where username was not properly pulled from
the principle object.

Changed formatDateTimePeriod to accept a style class rather than forcing a style
to be applied. This fixes the listActivity template so that the date occurred
appears in black text instead of gray, which is used on the activity template.