Consolidated all logic used to create and pre-fill a new ApplicationRelease into the ApplicationRelease service.
Fixed logic used to determine the next release number so that it checks for
existing releases of the application being released instead of all application
releases.

Prefill the buildPath of a new release with the buildPath from the last release
created.

When creating the first release of an application pre-fill release number with
1.0.0.