Replaced customized default Grails style with Twitter Bootstrap.
This is a wholesale replacement of the customized version of the Grails look and
feel with the Twitter Bootstrap framework plugin. This commit updates all GSP
pages to use the new look.

All forms now use a vertical layout with top labels.
Buttons are now left aligned with the primary button on the far left.
Easytabs jQuery plugin replaced with the bootstrap tabs code.
Badges were added to table text.
Minor issues/bugs found throughout the code have been addressed.

There is still work left to do. The application header/menu is not properly
styled. Quick search and user info dropdown are not functional.