Catch and log errors when sending email notifications
To prevent notifications from not being sent due to one bad email
address the notification service was modified so that it will catch and
log any exceptions that are caught when sending a notification email to
an email address. The notification service will continue to send email
messages to all other email addresses even when an error occurs. This is
a safe-gaurd against invalid email addresses.